requirements.txt
aiofiles==24.1.0
alembic==1.14.0
click==8.1.8
Flask==3.1.0
Flask_Cors==5.0.0
Flask_Migrate==4.1.0
flask_sqlalchemy==3.1.1
flask_wtf==1.2.2
injector==0.22.0
langchain==0.3.17
langchain_community==0.3.16
langchain_core==0.3.33
langchain_huggingface==0.1.2
langchain_openai==0.3.3
langchain_pinecone==0.2.2
langchain_text_splitters==0.3.5
langchain_weaviate==0.0.4
numpy==2.2.2
openai==1.61.0
pydantic==2.10.6
pytest==8.3.4
python-dotenv==1.0.1
transformers==4.48.1
weaviate_client==4.10.4
WTForms==3.2.1
